requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
config==0.5.1
Django==4.2.7
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
idna==3.4
Pillow==10.1.0
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
python-gettext==5.0
requests==2.31.0
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
